requirements.txt
pygame-ce
numpy
random
